select-fields: remove support for composite primary keys
phpstan reported this:
Call to function is_array() with string will always evaluate to false

After adding types to `getPrimaryKeyFromParentType()`, it's not possible
to return an array here.

All the information I could find on such composite keys is that Laravel
does not support them (e.g. https://github.com/laravel/framework/issues/5355 ).

I also read a bit through the 5.8 sources and everywhere where
`getKeyName()` is called within Laravel you can see that it's expected
to be a string and doesn't seem like it would truly work with an array.