lumen: fix path multi-level support
The existing code didn't work when accessing `/graphql/some/schema`,
only for `/graphql/some`.

With the current routing setup, the parameters to the route describe
the schema to use.

The existing code did check `request->count()` which is something else,
it contains e.g. query or JSON parameters and didn't return a correct
schema.

Route parameters are actually available in `$route[2]` which the code
now checks for.

This also means GraphiQL now finally fully works. The reason is that
GraphiQL sent multiple json parameters (operationName, query, variables)
which the code due to `request->all()` used to build the schema name
which never worked correctly.

It's not possible to write a test for this as orchestra/testbench
does not support Lumen currently.