Add support for per-operation query context
Previously the context was shared for all batch operations, but the
concept is that each operation executes indepdendently and thus each
has its own context.

It would technically still be possible to share a context by overriding
`queryContext` and come up with a custom solution.

This adds some small refactoring so that a query execution is
better encpasulated.