Fix batch request combined with file uploads
`->all()` returns the input _as well_ as the files in single array.

A non-batch request has the input keys `query` and optional
`operationName` and `variables`.

Even when `->all()` is used with non-batch and file uploads, the files
are inter-mixed with the resulting arrays using array offsets 0, 1, …

So a resulting input would look like:
```
[
  'query' => …,
  'variables' => …,
  '0' => uploaded file,
]
```

This is actually undesired but had no-side effects for non-batch
requests, as no one was looking at/accessing the 0 offset for example.

(Uploaded files are transplanted _within_ the `variables` anyway).

Now with batch requests, a conflict arises because the input itself is
an array (of queries / operations) _as well_ as the files:
```
[
  '0' => [
    'query' => …,
    'variables' => …,
  ],
  '0' => uploaded file,
```

The end result being that the files overwrite the batch operations and
thus completely breaking "batch queries plus file uploads".

The solution is to _just_ process the `->input()` for GraphQL; the
files are never needed directly here as the middleware transplants
them always within the `variables` within each batched request.