Make the register function same process idempotent
> Please also see the discussion in #508

It is rather simple really. 
The default testcase in lumen does an app refresh before each test in the setUp method, which does a bootstrap.
This means in the same process all classes are being re autoloaded, which is not a problem since it is done with require_once. But the GraphQL does not fall under that since is being re executed and not required. 

So there is an unmet requirement of the register function, it needs to be process idempotent.

I have modified my PR to make it be just that