Use testbench `WithConfig` attribute to set app.debug
This fixes the failing test `testNoExplicitContentType`

We do set it in `\Rebing\GraphQL\Tests\TestCase::getEnvironmentSetUp`
which _is_ executed in the this test, but too late.

By the time `\Illuminate\Foundation\Providers\FoundationServiceProvider::registerExceptionRenderer` is called,
the `getEnvironmentSetUp()` is not yet called and therefore the debug
mode being enabled is not realized.

But other parts in the framework later "see" the enabled debug mode,
and due to this mismatch the framework expects registered classes
which however arent.